#108 rename DatabaseManger to JdbcDatabaseManager and extract interface DatabaseManager from it
